last minor fixes and changes based upon PR feedback
