+++ tests/unit/test_general_configs.py
@@ -1,6 +1,8 @@
+from configs.general import _model_name
+from configs.general import _schematron_errors
 from configs.general import EicrConfig
-from configs.general import MODEL_NAME
-from configs.general import SCHEMATRON_ERRORS
+from configs.general import get_configuration_for_data_element
+from configs.general import get_data_element_from_schematron_error
 
 
 class TestGeneralConfigs:  # noqa: D101
@@ -24,13 +26,39 @@ class TestGeneralConfigs:  # noqa: D101
             assert isinstance(e, KeyError)
 
     def test_model_name(self):
-        assert MODEL_NAME is not None
-        assert isinstance(MODEL_NAME, str)
+        assert _model_name is not None
+        assert isinstance(_model_name, str)
 
     def test_schematron_errors(self):
-        assert len(SCHEMATRON_ERRORS) > 0
+        assert len(_schematron_errors) > 0
 
-        for df, errors in SCHEMATRON_ERRORS.items():
+        for df, errors in _schematron_errors.items():
             assert df in self.DATA_FIELDS
             assert isinstance(errors, list)
             assert len(errors) > 0
+
+    def test_get_data_field_config(self):
+        data_field = "lab_order"
+        config = get_configuration_for_data_element(data_field)
+
+        assert config is not None
+        assert hasattr(config, "schematron_errors")
+        assert hasattr(config, "text_word_count")
+        assert config.text_word_count == 2
+        assert len(config.xpaths) == 6
+
+    def test_get_data_field_config_wrong_data_element(self):
+        data_field = "MY_field"
+        config = get_configuration_for_data_element(data_field)
+
+        assert config is None
+
+    def test_get_data_field_by_schematron_error(self):
+        schematron_error = "Text to Code: Lab Test Name Ordered does not have a @code attribute"
+        data_field = get_data_element_from_schematron_error(schematron_error)
+
+        assert data_field == "lab_order"
+
+        schematron_error = "Some unknown error message"
+        data_field = get_data_element_from_schematron_error(schematron_error)
+        assert data_field is None
